+++ mvc/src/test/java/samples/TestHandlerAdapter.java
@@ -5,7 +5,7 @@ import jakarta.servlet.http.HttpServletResponse;
 
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
+import com.interface21.webmvc.servlet.mvc.HandlerAdapter;
 
 public class TestHandlerAdapter implements HandlerAdapter {
